+++ src/eca/llm_providers/openai.clj
@@ -428,10 +428,9 @@
 (defmethod f.login/login-step ["openai" :login/waiting-api-key] [{:keys [input db* provider send-msg!] :as ctx}]
   (if (string/starts-with? input "sk-")
     (do (config/update-global-config! {:providers {"openai" {:key input}}})
-        (swap! db* update :auth dissoc provider)
+        (swap! db* assoc-in [:auth provider] {:step :login/done :type :auth/token})
         (send-msg! (str "API key saved in " (.getCanonicalPath (config/global-config-file))))
-
-        (f.login/login-done! ctx :update-cache? false))
+        (f.login/login-done! ctx))
     (send-msg! (format "Invalid API key '%s'" input))))
 
 (defmethod f.login/login-step ["openai" :login/renew-token] [{:keys [db* provider] :as ctx}]